requirements.txt
pandas == 2.2.0
pyarrow == 17.0.0
python-dotenv == 1.0.1
matplotlib == 3.9.0
seaborn == 0.13.2
numpy == 2.1.1